+++ weed/storage/needle_map.go
@@ -8,6 +8,7 @@ import (
 	"github.com/seaweedfs/seaweedfs/weed/storage/idx"
 	"github.com/seaweedfs/seaweedfs/weed/storage/needle_map"
 	. "github.com/seaweedfs/seaweedfs/weed/storage/types"
+	"github.com/syndtr/goleveldb/leveldb/opt"
 )
 
 type NeedleMapKind int
@@ -41,6 +42,12 @@ type baseNeedleMapper struct {
 	indexFile           *os.File
 	indexFileAccessLock sync.Mutex
 	indexFileOffset     int64
+}
+
+type TempNeedleMapper interface {
+	NeedleMapper
+	DoOffsetLoading(v *Volume, indexFile *os.File, startFrom uint64) error
+	UpdateNeedleMap(v *Volume, indexFile *os.File, opts *opt.Options) error
 }
 
 func (nm *baseNeedleMapper) IndexFileSize() uint64 {
